double -> float

+++ src/MineCase.Algorithm/Noise/INoise.cs
@@ -7,8 +7,8 @@ namespace MineCase.Algorithm.Noise
 {
     public interface INoise
     {
-        double Noise(double x, double y, double z);
+        float Noise(float x, float y, float z);
 
-        void Noise(double[,,] noise, Vector3 offset, Vector3 scale);
+        void Noise(float[,,] noise, Vector3 offset, Vector3 scale);
     }
 }
